+++ sphinxcontrib/needs/defaults.py
@@ -1,4 +1,5 @@
 # flake8: noqa
+import os
 
 from docutils.parsers.rst import directives
 
@@ -216,3 +217,5 @@ NEED_DEFAULT_OPTIONS = {
 }
 
 NEEDEXTEND_NOT_ALLOWED_OPTIONS = ["id"]
+
+NEEDS_PROFILING = [x.upper() for x in os.environ.get("NEEDS_PROFILING", "").split(",")]