+++ sphinxcontrib/needs/directives/need.py
@@ -16,6 +16,7 @@ from sphinxcontrib.needs.functions import (
 from sphinxcontrib.needs.functions.functions import check_and_get_content
 from sphinxcontrib.needs.layout import build_need
 from sphinxcontrib.needs.logging import get_logger
+from sphinxcontrib.needs.utils import profile
 
 logger = get_logger(__name__)
 
@@ -268,6 +269,7 @@ def add_sections(app, doctree, fromdocname):
         #     needs[parent_need_id]["child_needs"].append(need_info["id"])
 
 
+@profile("NEED_PROCESS")
 def process_need_nodes(app, doctree, fromdocname):
     """
     Event handler to add title meta data (status, tags, links, ...) information to the Need node.
@@ -299,6 +301,7 @@ def process_need_nodes(app, doctree, fromdocname):
         create_back_links(env, links["option"])
 
 
+@profile("NEED_PRINT")
 def print_need_nodes(app, doctree, fromdocname):
     """
     Finally creates the need-node in the docurils node-tree.